pyproject.toml
[project]
name = "timesfm"
version = "2.0.0"
description = "A time series foundation model."
authors = [
    {name = "Rajat Sen", email = "senrajat@google.com"},
    {name = "Yichen Zhou", email = "yichenzhou@google.com"},
    {name = "Abhimanyu Das", email = "abhidas@google.com"},
    {name = "Petros Mol", email = "pmol@google.com"},
    {name = "Michael Chertushkin", email = "chertushkinmichael@gmail.com"},
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "huggingface_hub[cli]>=0.23.0",
    "safetensors>=0.5.3",
    "torch>=2.0.0",
    "fastapi[standard]>=0.120.0",
    "uvicorn[standard]>=0.38.0",
    "fastapi-cache2[redis]>=0.2.2",
]


[project.optional-dependencies]
torch = [
    "torch>=2.0.0",
]
flax = [
    "flax",
    "optax",
    "einshape",
    "orbax-checkpoint",
    "jaxtyping",
    "jax[cuda]"
]

[tool.ruff]
line-length = 88
indent-width = 2

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

